backend/requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi-mail==1.4.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
python-jose[cryptography]
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
passlib[bcrypt]

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
sqlalchemy-utils>=0.41.0
alembic>=1.12.0

# PostgreSQL dependencies
psycopg2-binary>=2.9.5
asyncpg>=0.27.0

# Redis dependencies
redis>=4.5.1

# Background job processing
celery[redis]>=5.3.0
kombu>=5.3.0

# Rate limiting
slowapi>=0.1.9
limits>=5.5.0

# HTTP client for Authentik integration
httpx>=0.25.0

# System monitoring
psutil>=5.9.0
